BDAG/Scripts/requirements.txt
affine==2.4.0
attrs==23.1.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pip==22.0.2
pyparsing==3.0.9
pypng==0.20220715.0
pyproj==3.5.0
python-dateutil==2.8.2
rasterio==1.3.6
setuptools==59.6.0
six==1.16.0
snuggs==1.4.7
